fix typing

+++ panda_gym/envs/tasks/flip.py
@@ -88,4 +88,4 @@ class Flip(Task):
         if self.reward_type == "sparse":
             return -np.array(d > self.distance_threshold, dtype=np.float32)
         else:
-            return -d
+            return -d.astype(np.float32)
