+++ panda_gym/envs/tasks/push.py
@@ -95,4 +95,4 @@ class Push(Task):
         if self.reward_type == "sparse":
             return -np.array(d > self.distance_threshold, dtype=np.float32)
         else:
-            return -d
+            return -d.astype(np.float32)
